Adds endpoint to fetch a candidate by Id

+++ src/main/java/com/evoter/repository/CandidateRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * @author showunmioludotun
  */
-public interface CandidateRepository extends JpaRepository<Candidate, Integer> {
+public interface CandidateRepository extends JpaRepository<Candidate, Long> {
 }
